ElasticSearchDestination-v2: ESMessageProcessor renamed to ESNativeMessageProcessor
